+++ foxes/models/turbine_models/sector_management.py
@@ -21,9 +21,15 @@ class SectorManagement(TurbineModel):
     target_vars : list of str
         The variables that change if range variables
         are within specified ranges
+    col_tinds : str, optional
+        The turbine index column name in the data
+    col_tnames : str, optional
+        The turbine name column name in the data
     colmap : dict
         Mapping from expected to existing
         column names
+    var_periods : dict
+        Periods for periodic variables
     pd_file_read_pars : dict, optional
         Parameters for pandas file reading
 
@@ -39,20 +45,22 @@ class SectorManagement(TurbineModel):
             data_source,
             range_vars,
             target_vars,
-            col_index=None,
-            col_turbine=None,
+            col_tinds=None,
+            col_tnames=None,
             colmap={},
+            var_periods={FV.WD: 360., FV.AMB_WD: 360.},
             pd_file_read_pars={},
         ):
         super().__init__()
 
         self.source = data_source
 
-        self._col_i = col_index
-        self._col_t = col_turbine
+        self._col_i = col_tinds
+        self._col_t = col_tnames
         self._rvars = range_vars
         self._tvars = target_vars
         self._colmap = colmap
+        self._perds = var_periods
         self._rpars = pd_file_read_pars
 
         self._rdata = None
@@ -81,19 +89,18 @@ class SectorManagement(TurbineModel):
                 if verbosity > 0:
                     print(f"{self.name}: Reading file {self.source}")
                 data = PandasFileHelper.read_file(self.source, **self._rpars)
-            
-            data.reset_index(inplace=True)
+
             if self._col_i is not None and self._col_t is None:
-                data.set_index(self._col_i, inplace=True)
+                data.reset_index(inplace=True)
             elif self._col_i is None and self._col_t is not None:
                 tnames = algo.farm.turbine_names()
                 inds = [tnames.index(name) for name in data[self._col_t]]
-                data[FV.turbine] = inds
-                data.set_index(FV.TURBINE, inplace=True)
+                data[FV.TURBINE] = inds
+                self._col_i = FV.TURBINE
             else:
-                raise KeyError(f"{self.name}: Please either specify 'col_index' or 'col_turbine'")
-            data.sort_index(inplace=True)
-            self._turbines = data.index.to_numpy()
+                raise KeyError(f"{self.name}: Please either specify 'col_tinds' or 'col_tnames'")
+            self._trbs = data[self._col_i].to_numpy()
+            n_trbs = len(self._trbs)
 
             self._rcols = []
             for v in self._rvars:
@@ -118,8 +125,12 @@ class SectorManagement(TurbineModel):
                 self._tcols.append(col)
 
             n_rvars = len(self._rvars)
-            self._rdata = data[self._rcols].to_numpy().reshape(n_rvars, 2)
+            self._rdata = data[self._rcols].to_numpy().reshape(n_trbs, n_rvars, 2)
             self._tdata = data[self._tcols].to_numpy()
+
+            for vi, v in enumerate(self._rvars):
+                if v in self._perds:
+                    self._rdata[:, vi] = np.mod(self._rdata[:, vi], self._perds[v])
 
         super().initialize(algo, st_sel, verbosity)
 
@@ -168,20 +179,33 @@ class SectorManagement(TurbineModel):
         """
 
         # prepare:
-        n_trbs = len(self._turbines)
-        tsel = np.s_[:] if np.all(self._turbines == np.arange(fdata.n_turbines)) else self._turbines
+        n_trbs = len(self._trbs)
+        if n_trbs == fdata.n_turbines and np.all(self._trbs == np.arange(fdata.n_turbines)):
+            tsel = np.s_[:]
+        else:
+            tsel = self._trbs
 
         # find state-turbine data that matches ranges:
         rsel = np.ones((fdata.n_states, n_trbs), dtype=bool)
         for vi, v in enumerate(self._rvars):
             d = fdata[v][:, tsel]
-            rsel = rsel & (d >= self._rdata[vi, 0]) & (d < self._rdata[vi, 1]) 
+            if v in self._perds:
+                d = np.mod(d, self._perds[v])
+                mi = self._rdata[:, vi, 0]
+                ma = self._rdata[:, vi, 1]
+                sel = ma < mi
+                if np.any(sel):
+                    rsel[:, sel] = rsel[:, sel] & ( (d[:, sel] >= mi[sel]) | (d[:, sel] < ma[sel]) ) 
+                if np.any(~sel):
+                    rsel[:, ~sel] = rsel[:, ~sel] & (d[:, ~sel] >= mi[~sel]) & (d[:, ~sel] < ma[~sel])
+            else:
+                rsel = rsel & (d >= self._rdata[None, :, vi, 0]) & (d < self._rdata[None, :, vi, 1]) 
 
         # set target data:
         if np.any(rsel):
-            sel = np.ones((fdata.n_states, algo.n_turbines), dtype=bool)
-            sel[: tsel] = rsel
+            sel = np.where(rsel)
+            selt = self._trbs[sel[1]]
             for vi, v in enumerate(self._tvars):
-                fdata[v][sel] = self._tdata[vi]
+                fdata[v][sel[0], selt] = self._tdata[None, sel[1], vi]
 
         return {v: fdata[v] for v in self._tvars}