+++ docs/source/history.md
@@ -122,4 +122,9 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.2.5-alpha
 
+- Turbine models:
+  - New: `SectorManagement`, sets variables by range rules on other variables
+- Examples:
+  - New: `sector_management`, demonstrates how to model wind sector management
+
 **Full Changelog**: [](https://github.com/FraunhoferIWES/foxes/commits/v0.2.5)